move one support method down to the generator using it
