+++ generators/base-application/support/entity-utils.ts
@@ -1,4 +1,12 @@
-import type { BaseApplicationEntity } from '../types.js';
+import type { BaseApplicationEntity, BaseApplicationField, BaseApplicationRelationship } from '../types.js';
 
-export type GetRelationshipType<E extends BaseApplicationEntity> = E extends { relationships: (infer R)[] } ? R : never;
-export type GetFieldType<E extends BaseApplicationEntity> = E extends { fields: (infer F)[] } ? F : never;
+export type GetRelationshipType<
+  F extends BaseApplicationField,
+  R extends BaseApplicationRelationship,
+  E extends BaseApplicationEntity<F, R>,
+> = E extends { relationships: (infer R)[] } ? R : never;
+export type GetFieldType<
+  F extends BaseApplicationField,
+  R extends BaseApplicationRelationship,
+  E extends BaseApplicationEntity<F, R>,
+> = E extends { fields: (infer F)[] } ? F : never;
