+++ generators/base-application/support/index.ts
@@ -24,8 +24,6 @@ export * from './enum.js';
 export * from './entity.js';
 export * from './entities.js';
 export * from './field-utils.js';
-export { default as prepareEntity } from './prepare-entity.js';
-export * from './prepare-entity.js';
 export { default as prepareField } from './prepare-field.js';
 export * from './prepare-field.js';
 export * from './prepare-property.js';
